+++ spec/unit/collect_spec.rb
@@ -1,14 +1,14 @@
 # encoding: utf-8
 
-RSpec.describe TTY::Prompt, '#gather' do
+RSpec.describe TTY::Prompt, '#collect' do
 
   subject(:prompt) { TTY::TestPrompt.new }
 
-  it "gathers more than one answer" do
+  it "collects more than one answer" do
     prompt.input << "Piotr\r30\rStreet\rCity\r123\r"
     prompt.input.rewind
 
-    result = prompt.gather do
+    result = prompt.collect do
       key(:name).ask('Name?')
 
       key(:age).ask('Age?', convert: :int)